requirements.txt
email_validator
flask
flask-admin
flask-cors
flask-login
flask-restful
flask-socketio
flask-sqlalchemy
flask-wtf
eventlet
python-chess
tabulate
